requirements.txt
pyyaml
pandas
matplotlib
streamlit
